requirements.txt
streamlit
tensorflow
numpy
opencv-python-headless
pillow
matplotlib
gtts
